src/backend/pyproject.toml
[project]
name = "Bot_ferme"
version = "0.1.0"
description = ""
authors = [
    { name = "Ivan Gerasimov", email = "ivan.gerasimou@yandex.ru" },
]
license = { text = "MIT" }
readme = "../../README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.121.1,<0.122.0",
    "uvicorn[standard]>=0.38.0,<0.39.0",
    "sqlalchemy[asyncio]>=2.0.44,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "alembic>=1.17.1,<2.0.0",
    "pydantic[email]>=2.12.4,<3.0.0",
    "gunicorn>=23.0.0,<24.0.0",
    "inflect>=7.5.0,<8.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "bcrypt>=5.0.0,<6.0.0",
    "pytz>=2025.2,<2026.0",
    "pyjwt>=2.10.1,<3.0.0",
    "cryptography>=46.0.6",
    "pre-commit>=4.5.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.4.0,<5.0.0",
    "ruff>=0.14.4,<0.15.0",
    "black>=25.9.0,<26.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.18.2,<2.0.0",
]

[tool.uv]
managed = true
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
